pyproject.toml
[tool.poetry]
name = "tayra"
version = "0.1.0-alpha"
description = ""
authors = ["Filipe Scoton <filipemo@microsoft.com>", "Karina Assini <karinaa@microsoft.com>", "Ricardo Cataldi <rcataldi@microsoft.com>", "Luiz Braz <lfbraz@microsoft.com>"]
readme = "README.md"
packages = [{include = "src/."}]


[tool.poetry.dependencies]
python = "^3.12"
tiktoken = "^0.7.0"
fastapi = "^0.112.0"
azure-storage-blob = "^12.22.0"
httpx = "^0.27.0"
python-multipart = "^0.0.9"
aiohttp = "^3.10.3"
uvicorn = "^0.30.6"
python-dotenv = "^1.0.1"
azure-cosmos = {version = "^4.7.0", allow-prereleases = true}
aistudio-request = {git = "https://github.com/Azure-Samples/ai-requests-template.git"}
numpy = "^2.1.1"
scipy = "^1.14.1"
promptflow = {extras = ["all"], version = "^1.16.1"}


[tool.poetry.group.dev.dependencies]
pylint = "^2.15.9"
black = {extras = ["jupyter"], version = "^24.4.2"}
pytest = "^7.4.3"
isort = "^5.13.2"
pytest-asyncio = "^0.23.6"
ipykernel = "^6.29.3"
poetry-plugin-export = "^1.7.1"
locust = "^2.31.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q -s"
testpaths = ["tests",]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = ["async: marks tests as async (select with '-m async')", "slow: marks tests as slow (select with '-m slow')"]


[tool.pylint]
fail-under = 8.5
max-line-length = 100
max-complexity=10
good-names = ["i", "j", "k", "ex", "Run", "_"]
disable = ["C0114", "C0115", "C0116", "C0301", "W3101"]
load-plugins = ["pylint.extensions.mccabe", ]


[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'
exclude = '(\.git | \.hg | \.mypy_cache | \.tox | \.venv | _build | buck-out | build | dist )/'


[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 100